refactor: turn dataset selector input state inside out
So that it can be customized
